+++ IBAnimatable/AnimatableViewController.swift
@@ -50,7 +50,7 @@ import UIKit
     
     let toViewController = segue.destinationViewController
     // If interactiveGestureType has been set
-    if let interactiveGestureType = interactiveGestureType, interactiveGestureTypeValue = InteractiveGestureType(rawValue: interactiveGestureType) {
+    if let interactiveGestureType = interactiveGestureType, interactiveGestureTypeValue = InteractiveGestureType.fromString(interactiveGestureType) {
       toViewController.transitioningDelegate = PresenterManager.sharedManager().retrievePresenter(animationType, transitionDuration: transitionDuration, interactiveGestureType: interactiveGestureTypeValue)
     }
     else {