+++ IBAnimatable/Constants.swift
@@ -7,4 +7,4 @@ import Foundation
 
 // Default transition duration.
 // The default UIKit transtion animation duration is 0.35, but too fast for most of custom transition animations.
-let defaultTransitionDuration: Duration = 0.5
+let defaultTransitionDuration: Duration = 0.75