+++ IBAnimatable/TransitionAnimationType.swift
@@ -14,10 +14,15 @@ public enum TransitionAnimationType {
   case FadeOut          // FromView Fades out
   case SystemSuckEffect
   case SystemRippleEffect
+  case Explode(params: String)
   case SystemCube(direction: TransitionFromDirection)
   case SystemFlip(direction: TransitionFromDirection)
+  case SystemMoveIn(direction: TransitionFromDirection)
+  case SystemPush(direction: TransitionFromDirection)
+  case SystemReveal(direction: TransitionFromDirection)
   case SystemPage(type: TransitionPageType)
   case SystemCameraIris(hollowState: TransitionHollowState)
+  case SystemRotate(degree: TransitionRotateDegree)
     
   var stringValue: String {
     return String(self)
@@ -28,12 +33,16 @@ public enum TransitionAnimationType {
       return .SystemRippleEffect
     } else if transitionType.hasPrefix("SystemSuckEffect") {
         return .SystemSuckEffect
+    } else if transitionType.hasPrefix("Explode") {
+      return .Explode(params: params(forTransitionType: transitionType))
     } else if transitionType.hasPrefix("Fade") {
       return fadeTransitionAnimationType(transitionType)
     } else if transitionType.hasPrefix("SystemCameraIris") {
         return cameraIrisTransitionAnimationType(transitionType)
     } else if transitionType.hasPrefix("SystemPage") {
       return pageTransitionAnimationType(transitionType)
+    } else if transitionType.hasPrefix("SystemRotate") {
+      return rotateTransitionAnimationType(transitionType)
     } else  {
       return fromStringWithDirection(transitionType)
     }
@@ -55,7 +64,7 @@ private extension TransitionAnimationType {
   }
  
   static func cameraIrisTransitionAnimationType(transitionType: String) -> TransitionAnimationType? {
-    let transitionType = cleanTransitionType(transitionType)
+    let transitionType = params(forTransitionType: transitionType)
     if transitionType.containsString("hollowopen") {
       return .SystemCameraIris(hollowState: .Open)
     } else if transitionType.containsString("hollowclose") {
@@ -65,11 +74,25 @@ private extension TransitionAnimationType {
   }
   
   static func pageTransitionAnimationType(transitionType: String) -> TransitionAnimationType? {
-    let transitionType = cleanTransitionType(transitionType)
+    let transitionType = params(forTransitionType: transitionType)
     if transitionType.containsString("uncurl") {
       return .SystemPage(type: .UnCurl)
     } else if transitionType.containsString("curl") {
       return .SystemPage(type: .Curl)      
+    }
+    return nil
+  }
+  
+  static func rotateTransitionAnimationType(transitionType: String) -> TransitionAnimationType? {
+    let transitionType = params(forTransitionType: transitionType)
+    if transitionType.containsString("90ccw") || transitionType.containsString("ninetyccw") {
+      return .SystemRotate(degree: .NinetyCCW)
+    } else if transitionType.containsString("90") || transitionType.containsString("ninety") {
+      return .SystemRotate(degree: .Ninety)
+    } else if transitionType.containsString("180ccw") || transitionType.containsString("OneHundredHeightyccw") {
+      return .SystemRotate(degree: .OneHundredHeightyCCW)
+    } else if transitionType.containsString("180") || transitionType.containsString("OneHundredHeighty") {
+      return .SystemRotate(degree: .OneHundredHeighty)
     }
     return nil
   }
@@ -83,6 +106,12 @@ private extension TransitionAnimationType {
       return .SystemCube(direction: direction)
     } else if transitionType.hasPrefix("SystemFlip") {
       return .SystemFlip(direction: direction)
+    } else if transitionType.hasPrefix("SystemMoveIn") {
+      return .SystemMoveIn(direction: direction)
+    } else if transitionType.hasPrefix("SystemPush") {
+      return .SystemPush(direction: direction)
+    } else if transitionType.hasPrefix("SystemReveal") {
+      return .SystemReveal(direction: direction)
     }
     return nil
   }
@@ -93,16 +122,18 @@ private extension TransitionAnimationType {
 
 private extension TransitionAnimationType {
   
-  static func cleanTransitionType(transitionType: String) -> String {
+  static func params(forTransitionType transitionType: String) -> String {
     let range = transitionType.rangeOfString("(")
-    let transitionType = transitionType.stringByReplacingOccurrencesOfString(" ", withString: "")
+    var transitionType = transitionType.stringByReplacingOccurrencesOfString(" ", withString: "")
       .lowercaseString
       .substringFromIndex(range?.startIndex ?? transitionType.endIndex)
+      .stringByReplacingOccurrencesOfString("(", withString: "")
+      .stringByReplacingOccurrencesOfString(")", withString: "")
     return transitionType
   }
   
   static func transitionDirection(forTransitionType transitionType: String) -> TransitionFromDirection? {
-    let transitionType = cleanTransitionType(transitionType)
+    let transitionType = params(forTransitionType: transitionType)
     if transitionType.containsString("left") {
       return .Left
     } else if transitionType.containsString("right") {